+++ client/src/views/login/Login.vue
@@ -2,7 +2,8 @@
     <v-container class="fill-height d-flex align-center">
         <v-row justify="center">
             <v-card
-                min-width="320px"
+                :min-width="320"
+                :max-width="500"
                 elevation="10"
                 rounded="xl"
                 class="pa-7 pb-3"
@@ -11,7 +12,10 @@
                     {{ appStore.state.environmentVariables.appTitle }}
                 </v-card-title>
 
-                <v-tabs-window v-model="currentTab">
+                <v-tabs-window
+                    v-model="currentTab"
+                    @update:model-value="onTabChange"
+                >
                     <!-- 登入 -->
                     <v-tabs-window-item value="login">
                         <v-card-text>
@@ -161,13 +165,80 @@
                             </v-btn>
                         </v-card-text>
                     </v-tabs-window-item>
+
+                    <!-- 驗證碼 -->
+                    <v-tabs-window-item value="verification_code">
+                        <v-card-text>
+                            <v-form validate-on="input">
+                                <v-row>
+                                    <!-- Email -->
+                                    <v-col cols="12">
+                                        <p class="text-caption">
+                                            {{ t('login.input_label_email') }}
+                                        </p>
+                                        <v-text-field
+                                            v-model="form.email.value.value"
+                                            :error-messages="form.email.errorMessage.value"
+                                            autocomplete="username"
+                                            variant="underlined"
+                                            hide-details="auto"
+                                            :placeholder="t('login.input_label_email')"
+                                            autofocus
+                                            required
+                                        />
+                                    </v-col>
+                                    <!-- 驗證碼 -->
+                                    <v-col cols="12">
+                                        <p class="text-caption">
+                                            {{ t('login.label_verification_code') }}
+                                        </p>
+                                        <v-text-field
+                                            v-model="form.verificationCode.value.value"
+                                            :error-messages="form.verificationCode.errorMessage.value"
+                                            hide-details="auto"
+                                            :placeholder="t('login.input_label_verification_code')"
+                                            required
+                                            variant="underlined"
+                                        />
+                                    </v-col>
+                                </v-row>
+                            </v-form>
+                        </v-card-text>
+
+                        <v-card-text>
+                            <v-btn
+                                color="warning"
+                                flat
+                                rounded="xl"
+                                class="mb-3"
+                                :loading="loading"
+                                block
+                                @click="onResendVerificationCodeClick"
+                            >
+                                {{ t('login.button_resend_verification_code') }}
+                            </v-btn>
+                            <v-btn
+                                color="primary"
+                                :disabled="!form.verificationCode"
+                                flat
+                                rounded="xl"
+                                :loading="loading"
+                                block
+                                @click="onSubmit"
+                            >
+                                {{ t('button.submit') + t('login.label_verification_code') }}
+                            </v-btn>
+                        </v-card-text>
+                    </v-tabs-window-item>
                 </v-tabs-window>
-
-                <v-divider>或</v-divider>
-
-                <v-card-text>
-                    <GoogleLogin @update:credential="onGoogleLogin" />
-                </v-card-text>
+              
+                <template v-if="currentTab !== 'verification_code'">
+                    <v-divider>或</v-divider>
+                    
+                    <v-card-text>
+                        <GoogleLogin @update:credential="onGoogleLogin" />
+                    </v-card-text>
+                </template>
 
                 <v-tabs
                     v-model="currentTab"
@@ -176,7 +247,7 @@
                     hide-slider
                 >
                     <v-tab
-                        v-if="currentTab === 'login'"
+                        v-if="currentTab !== 'signup'"
                         value="signup"
                         class="text-info"
                         variant="plain"
@@ -185,13 +256,22 @@
                         {{ t('login.button_signup') }}
                     </v-tab>
                     <v-tab
-                        v-if="currentTab === 'signup'"
+                        v-if="currentTab !== 'login'"
                         value="login"
                         class="text-info"
                         variant="plain"
                         :ripple="false"
                     >
                         {{ t('login.button_login') }}
+                    </v-tab>
+                    <v-tab
+                        v-if="currentTab !== 'verification_code'"
+                        value="verification_code"
+                        class="text-info"
+                        variant="plain"
+                        :ripple="false"
+                    >
+                        {{ t('login.button_resend_verification_code') }}
                     </v-tab>
                 </v-tabs>
             </v-card>
@@ -222,6 +302,7 @@ import { useAppStore } from '@/store/app';
 import { useAuthStore } from '@/store/auth';
 import { useNotifierStore } from '@/store/notifier';
 import { Auth } from '@/types/auth';
+import { REGEX_PASSWORD } from '@/utils/credential';
 
 import GoogleLogin from './components/GoogleLogin.vue';
 
@@ -233,38 +314,60 @@ const route = useRoute();
 const showPassword = ref(false);
 const appStore = useAppStore();
 
-const currentTab = ref<'login' | 'signup'>('login');
+const currentTab = ref<'login' | 'signup' | 'verification_code'>('login');
 
 const loading = ref(false);
+const isVerificationCodeSent = ref(false);
 
-const { handleSubmit } = useForm({
+const {
+    handleSubmit, resetField, 
+} = useForm({
     initialValues: {
         email: '',
         password: '',
         name: '',
         passwordConfirm: '',
+        verificationCode: '',
     },
     validationSchema: yup.object({
         email: yup
-            .string()
-            .required(t('input.error_required')),
-        password: yup
             .string()
             .required(t('input.error_required')),
         name: yup
             .string()
             .test('require', t('input.error_required'), (value) => {
                 if (currentTab.value === 'signup') {
-                    return value.length > 0;
+                    return value.trim().length > 0;
                 }
-
+                
                 return true;
             }),
+        password: yup
+            .string()
+            .required(t('input.error_required'))
+            .matches(REGEX_PASSWORD, '密碼必須由 6~10 位英數字組成'),
         passwordConfirm: yup
             .string()
             .test('require', t('input.error_required'), (value) => {
                 if (currentTab.value === 'signup') {
-                    return value.length > 0;
+                    return value.trim().length > 0;
+                }
+
+                return true;
+            })
+            .test('must-same', '需與密碼一致', (value, ctx) => {
+                if (currentTab.value === 'signup') {
+
+                    return ctx.parent?.password && ctx.parent.password === value;
+                }
+
+                return true;
+            }),
+        verificationCode: yup
+            .string()
+            .test('require', t('input.error_required'), (value) => {
+                if (currentTab.value === 'verification_code') {
+                    return value.trim().length > 0;
                 }
 
                 return true;
@@ -277,6 +380,7 @@ const form = {
     password: useField<string>('password'),
     name: useField<string>('name'),
     passwordConfirm: useField<string>('passwordConfirm'),
+    verificationCode: useField<string>('verificationCode'),
 };
 
 const doLogin = async (data: Auth.Login.Request) => {
@@ -293,6 +397,24 @@ const doLogin = async (data: Auth.Login.Request) => {
     }
 };
 
+const onResendVerificationCodeClick = async () => {
+    loading.value = true;
+    const params: Auth.SendVerificationCode.Request = { email: form.email.value.value };
+
+    try {
+        await AuthService.sendVerificationCode(params);
+
+        notifierStore.success({ content: t('login.message_send_verification_code_success') });
+
+        isVerificationCodeSent.value = true;
+    } catch (err) {
+        console.error(err);
+        notifierStore.error({ content: t('login.message_send_verification_code_fail') });
+    }
+
+    loading.value = false;
+};
+
 const onSubmit = handleSubmit(async (value) => {
     loading.value = true;
 
@@ -304,7 +426,9 @@ const onSubmit = handleSubmit(async (value) => {
         };
 
         await doLogin(params);
-    } else if (currentTab.value === 'signup') {
+    }
+    // 註冊
+    else if (currentTab.value === 'signup') {
         const params: Auth.SignUp.Request = {
             email: value.email,
             name: value.name,
@@ -316,10 +440,29 @@ const onSubmit = handleSubmit(async (value) => {
     
             notifierStore.success({ content: t('login.message_signup_success') });
 
-            currentTab.value = 'login';
+            isVerificationCodeSent.value = true;
+            currentTab.value = 'verification_code';
         } catch (err) {
             console.error(err);
             notifierStore.error({ content: t('login.message_sugnup_fail') });
+        }
+    }
+    // 填驗證碼
+    else if (currentTab.value === 'verification_code') {
+        const params: Auth.VerifyCode.Request = {
+            email: value.email,
+            verificationCode: value.verificationCode,
+        };
+
+        try {
+            await AuthService.verifyCode(params);
+
+            notifierStore.success({ content: t('login.message_verify_account_success') });
+
+            currentTab.value = 'login';
+        } catch (err) {
+            console.error(err);
+            notifierStore.error({ content: t('login.message_verify_account_fail') });
         }
     }
 
@@ -333,6 +476,15 @@ const onGoogleLogin = async (code: string) => {
     };
 
     await doLogin(params);
+};
+
+const onTabChange = () => {
+    if (currentTab.value === 'verification_code') {
+        resetField('verificationCode');
+    } else if (currentTab.value === 'signup') {
+        resetField('email');
+        resetField('passwordConfirm');
+    }
 };
 
 onMounted(() => {