+++ server/src/modules/accounts/dto/update-account.dto.ts
@@ -1,4 +1,9 @@
-import { IsString } from 'class-validator';
+import {
+    IsNotEmpty,
+    IsString,
+    Matches,
+} from 'class-validator';
+import { REGEX_PASSWORD } from 'src/utils/credential';
 
 import {
     ApiProperty,
@@ -11,8 +16,10 @@ export class UpdateAccountDto {
     @ApiProperty()
         name: string;
 
+    @IsNotEmpty()
     @IsString()
-    @ApiProperty()
+    @Matches(REGEX_PASSWORD, { message: 'Must use only a-z or 0-9' })
+    @ApiProperty({ description: 'At least 6, most 10 characters' })
         password: string;
 }