+++ server/src/utils/credential.ts
@@ -2,6 +2,9 @@ import * as bcrypt from 'bcrypt';
 import { AccountAuth } from 'src/modules/accounts/entities/account-auth.entity';
 import { LoginOrCreateDto } from 'src/modules/auth/dto/login-or-create.dto';
 
+/** 密碼組成規則: 6~10 位英數字 */
+const REGEX_PASSWORD = /^(?=.*[a-zA-Z])(?=.*\d)[a-zA-Z\d]{6,10}$/;
+
 function hashPassword(password: string) {
     return bcrypt.hashSync(password, 10);
 }
@@ -23,6 +26,7 @@ function verifyPasswordLogin(inputAuth: LoginOrCreateDto, dbAuth: AccountAuth) {
 
 export {
     hashPassword,
+    REGEX_PASSWORD,
     verifyPassword,
     verifyPasswordLogin,
 };